pyproject.toml
[project]
name = "smolmodels"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets",
    "evaluate",
    "wandb",
    "fire",
    "tensorboard",
    "huggingface-hub[hf-transfer]",
    "tabulate==0.9.0",
    "python-dotenv==1.0.1",
    "fsspec==2023.10.0",
    "matplotlib",
    "openai==1.74.0",
    "unidecode",
    "anthropic==0.49.0",
    "pillow==10.4.0",
    "loguru",
    "webdataset",
    "wheel",
    "datacommons",
    "datacommons-pandas>=0.0.3",
    "ipykernel>=6.29.5",
    "rich>=13.9.2",
    "polars>=1.12.0",
    "wrapt-timeout-decorator>=1.5.1",
    "ninja>=1.11.1.1",
    "packaging>=24.1",
    "fastapi>=0.115.5",
    "uvicorn>=0.32.1",
    "gradio==5.27.0",
    "scipy",
    "pyright>=1.1.390",
    "google-genai==1.12.1",
    "sentencepiece",
    "einops",
    "mistune>=3.1.0",
    "ruff>=0.8.1",
    "chardet>=5.2.0",
    "jsonlines>=4.0.0",
    "seaborn>=0.13.2",
    "plotting>=0.0.7",
    "pandas>=2.2.3",
    "modal==0.73.135",
    "pydantic>=2.10.6",
    "tiktoken",
    "sigstore>=1.0.0",
    "kagglehub>=0.3.7",
    "nltk>=3.9.1",
    "pytest>=8.3.4",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
torch = [
    "transformers==4.51.2",
    "torch==2.5.1",
    "torchaudio==2.5.1",
    "torchvision==0.20.1",
    "peft>=0.13.2",
    "trl==0.16.1",
    "lightning>=2.5.0.post0",
    "nltk",
]

training = [
    "setuptools_scm",
    "cmake",
    "bitsandbytes>=0.44.1; sys_platform == 'linux'",
    "liger-kernel>=0.5.2; sys_platform == 'linux'",
    "flash-attn; sys_platform == 'linux'",
    "vllm==0.7.3",
]

[tool.uv]
default-groups = ["torch"]


[tool.uv.sources]
torch = [{ index = "pytorch-cu124", marker = "sys_platform != 'darwin'" }]
torchvision = { index = "pytorch-cu124", marker = "sys_platform != 'darwin'" }
torchaudio = { index = "pytorch-cu124", marker = "sys_platform != 'darwin'" }


[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true


[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
